# openapi-formatted.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/audit-rules": {
            "post": {
                "operationId": "AuditRuleController_create",
                "summary": "\u521b\u5efa\u5ba1\u8ba1\u89c4\u5219",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAuditRuleDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "audit-rule"
                ]
            }
        },
        "/api/v1/audit-rules/list": {
            "get": {
                "operationId": "AuditRuleController_findAll",
                "summary": "\u83b7\u53d6\u5ba1\u8ba1\u89c4\u5219\u5217\u8868",
                "parameters": [
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "description": "\u9875\u7801",
                        "schema": {
                            "default": 1,
                            "example": 1,
                            "type": "number"
                        }
                    },
                    {
                        "name": "pageSize",
                        "required": false,
                        "in": "query",
                        "description": "\u6bcf\u9875\u6570\u91cf",
                        "schema": {
                            "default": 10,
                            "example": 10,
                            "type": "number"
                        }
                    },
                    {
                        "name": "keyword",
                        "required": false,
                        "in": "query",
                        "description": "\u5173\u952e\u8bcd\u641c\u7d22",
                        "schema": {
                            "example": "\u5408\u540c",
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "required": false,
                        "in": "query",
                        "description": "\u72b6\u6001\uff1a1-\u542f\u7528\uff0c0-\u505c\u7528",
                        "schema": {
                            "example": 1,
                            "enum": [
                                0,
                                1
                            ],
                            "type": "number"
                        }
                    },
                    {
                        "name": "auditType",
                        "required": false,
                        "in": "query",
                        "description": "\u5ba1\u8ba1\u7c7b\u578b",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phase",
                        "required": false,
                        "in": "query",
                        "description": "\u9636\u6bb5",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "verifySection",
                        "required": false,
                        "in": "query",
                        "description": "\u67e5\u8bc1\u677f\u5757",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "audit-rule"
                ]
            }
        },
...
        "/api/v1/audit-rules/export": {
            "get": {
                "operationId": "AuditRuleController_exportExcel",
                "summary": "\u5bfc\u51fa\u6570\u636e\u5230Excel",
                "parameters": [
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "description": "\u9875\u7801",
                        "schema": {
                            "default": 1,
                            "example": 1,
                            "type": "number"
                        }
                    },
                    {
                        "name": "pageSize",
                        "required": false,
                        "in": "query",
                        "description": "\u6bcf\u9875\u6570\u91cf",
                        "schema": {
                            "default": 10,
                            "example": 10,
                            "type": "number"
                        }
                    },
                    {
                        "name": "keyword",
                        "required": false,
                        "in": "query",
                        "description": "\u5173\u952e\u8bcd\u641c\u7d22",
                        "schema": {
                            "example": "\u5408\u540c",
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "required": false,
                        "in": "query",
                        "description": "\u72b6\u6001\uff1a1-\u542f\u7528\uff0c0-\u505c\u7528",
                        "schema": {
                            "example": 1,
                            "enum": [
                                0,
                                1
                            ],
                            "type": "number"
                        }
                    },
                    {
                        "name": "auditType",
                        "required": false,
                        "in": "query",
                        "description": "\u5ba1\u8ba1\u7c7b\u578b",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phase",
                        "required": false,
                        "in": "query",
                        "description": "\u9636\u6bb5",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "verifySection",
                        "required": false,
                        "in": "query",
                        "description": "\u67e5\u8bc1\u677f\u5757",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "audit-rule"
                ]
            }
        },
        "/api/v1/audit-rules/field-by-code/{fieldCode}": {
            "get": {
                "operationId": "AuditRuleController_getFieldByCode",
                "summary": "\u6839\u636e\u5b57\u6bb5\u7f16\u7801\u83b7\u53d6\u5b57\u6bb5\u4fe1\u606f",
                "parameters": [
                    {
                        "name": "fieldCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "audit-rule"
                ]
            }
        },
        "/api/v1/audit-rules/import": {
            "post": {
                "operationId": "AuditRuleController_importExcel",
                "summary": "Excel\u6279\u91cf\u5bfc\u5165",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "audit-rule"
                ]
            }
        },
        "/api/v1/audit-rules/{id}": {
            "get": {
                "operationId": "AuditRuleController_findOne",
                "summary": "\u83b7\u53d6\u5ba1\u8ba1\u89c4\u5219\u8be6\u60c5",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "audit-rule"
                ]
            },
            "put": {
                "operationId": "AuditRuleController_update",
                "summary": "\u66f4\u65b0\u5ba1\u8ba1\u89c4\u5219",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAuditRuleDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "audit-rule"
                ]
            },
            "delete": {
                "operationId": "AuditRuleController_remove",
                "summary": "\u5220\u9664/\u505c\u7528\u5ba1\u8ba1\u89c4\u5219",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "audit-rule"
                ]
            }
        },
        "/api/v1/audit-rules/batch/enable": {
            "post": {
                "operationId": "AuditRuleController_batchEnable",
                "summary": "\u6279\u91cf\u542f\u7528",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BatchIdsDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "audit-rule"
                ]
            }
        },
        "/api/v1/audit-rules/batch/disable": {
            "post": {
                "operationId": "AuditRuleController_batchDisable",
                "summary": "\u6279\u91cf\u505c\u7528",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BatchIdsDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "audit-rule"
                ]
            }
        },
        "/api/v1/audit-rules/batch/delete": {
            "post": {
                "operationId": "AuditRuleController_batchDelete",
                "summary": "\u6279\u91cf\u5220\u9664",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BatchIdsDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "audit-rule"
                ]
            }
        },
...
    },
    "info": {
        "title": "\u5de5\u7a0b\u54a8\u8be2\u5168\u4e1a\u52a1\u6570\u636e\u5e93\u5e73\u53f0 API",
        "description": "\n## \u6982\u8ff0\n\u5de5\u7a0b\u54a8\u8be2\u5168\u4e1a\u52a1\u6570\u636e\u5e93\u5e73\u53f0\u540e\u7aef\u63a5\u53e3\u6587\u6863 v1.0\n\n## \u8ba4\u8bc1\n\u6240\u6709\u63a5\u53e3\uff08\u9664\u767b\u5f55\u3001\u6ce8\u518c\u3001\u5065\u5eb7\u68c0\u67e5\u5916\uff09\u5747\u9700\u8981 JWT \u8ba4\u8bc1\u3002\n\u8bf7\u5728\u8bf7\u6c42\u5934\u4e2d\u6dfb\u52a0\uff1a`Authorization: Bearer <token>`\n\n## \u54cd\u5e94\u683c\u5f0f\n- \u6210\u529f\uff1a`{ data: T, meta?: { page, pageSize, total, totalPages } }`\n- \u5931\u8d25\uff1a`{ code: string, message: string, traceId: string, details?: any }`\n\n## \u5206\u9875\n- \u9ed8\u8ba4 page=1, pageSize=10\n- \u6700\u5927 pageSize=1000\n    ",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "JWT": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http",
                "description": "\u8f93\u5165 JWT Token \u8fdb\u884c\u8ba4\u8bc1"
            }
        },
        "schemas": {
...
            "BatchIdsDto": {
                "type": "object",
                "properties": {
                    "ids": {
                        "description": "ID\u6570\u7ec4",
                        "type": "array",
                        "items": {
                            "type": "number"
                        }
                    }
                },
                "required": [
                    "ids"
                ]
            },
...
            "CreateAuditRuleDto": {
                "type": "object",
                "properties": {
                    "ruleCode": {
                        "type": "string",
                        "description": "\u89c4\u5219\u7f16\u7801\uff08\u7559\u7a7a\u81ea\u52a8\u751f\u6210\uff09"
                    },
                    "ruleName": {
                        "type": "string",
                        "description": "\u89c4\u5219\u540d\u79f0"
                    },
                    "auditType": {
                        "type": "string",
                        "description": "\u5ba1\u8ba1\u7c7b\u578b"
                    },
                    "phase": {
                        "type": "string",
                        "description": "\u9636\u6bb5"
                    },
                    "verifySection": {
                        "type": "string",
                        "description": "\u67e5\u8bc1\u677f\u5757"
                    },
                    "problemDesc": {
                        "type": "string",
                        "description": "\u95ee\u9898\u63cf\u8ff0"
                    },
                    "compareMethod": {
                        "type": "string",
                        "description": "\u6bd4\u5bf9\u65b9\u5f0f"
                    },
                    "compareMethodLlm": {
                        "type": "string",
                        "description": "\u6bd4\u5bf9\u65b9\u5f0f-LLM\u7528"
                    },
                    "auditBasis": {
                        "type": "string",
                        "description": "\u5ba1\u8ba1\u4f9d\u636e\u5185\u5bb9"
                    },
                    "source1Code": {
                        "type": "string",
                        "description": "\u6570\u636e\u6e901\u7f16\u7801"
                    },
                    "source1Name": {
                        "type": "string",
                        "description": "\u6570\u636e\u6e901\u540d\u79f0"
                    },
                    "source2Code": {
                        "type": "string",
                        "description": "\u6570\u636e\u6e902\u7f16\u7801"
                    },
                    "source2Name": {
                        "type": "string",
                        "description": "\u6570\u636e\u6e902\u540d\u79f0"
                    },
                    "source3Code": {
                        "type": "string",
                        "description": "\u6570\u636e\u6e903\u7f16\u7801"
                    },
                    "source3Name": {
                        "type": "string",
                        "description": "\u6570\u636e\u6e903\u540d\u79f0"
                    },
                    "source4Code": {
                        "type": "string",
                        "description": "\u6570\u636e\u6e904\u7f16\u7801"
                    },
                    "source4Name": {
                        "type": "string",
                        "description": "\u6570\u636e\u6e904\u540d\u79f0"
                    },
                    "source5Code": {
                        "type": "string",
                        "description": "\u6570\u636e\u6e905\u7f16\u7801"
                    },
                    "source5Name": {
                        "type": "string",
                        "description": "\u6570\u636e\u6e905\u540d\u79f0"
                    },
                    "lawClauseCode": {
                        "type": "string",
                        "description": "\u6cd5\u6761\u7f16\u7801"
                    },
                    "remark": {
                        "type": "string",
                        "description": "\u5907\u6ce8"
                    }
                },
                "required": [
                    "ruleName"
                ]
            },
            "UpdateAuditRuleDto": {
                "type": "object",
                "properties": {
                    "ruleCode": {
                        "type": "string",
                        "description": "\u89c4\u5219\u7f16\u7801\uff08\u7559\u7a7a\u81ea\u52a8\u751f\u6210\uff09"
                    },
                    "ruleName": {
                        "type": "string",
                        "description": "\u89c4\u5219\u540d\u79f0"
                    },
                    "auditType": {
                        "type": "string",
                        "description": "\u5ba1\u8ba1\u7c7b\u578b"
                    },
                    "phase": {
                        "type": "string",
                        "description": "\u9636\u6bb5"
                    },
                    "verifySection": {
                        "type": "string",
                        "description": "\u67e5\u8bc1\u677f\u5757"
                    },
                    "problemDesc": {
                        "type": "string",
                        "description": "\u95ee\u9898\u63cf\u8ff0"
                    },
                    "compareMethod": {
                        "type": "string",
                        "description": "\u6bd4\u5bf9\u65b9\u5f0f"
                    },
                    "compareMethodLlm": {
                        "type": "string",
                        "description": "\u6bd4\u5bf9\u65b9\u5f0f-LLM\u7528"
                    },
                    "auditBasis": {
                        "type": "string",
                        "description": "\u5ba1\u8ba1\u4f9d\u636e\u5185\u5bb9"
                    },
                    "source1Code": {
                        "type": "string",
                        "description": "\u6570\u636e\u6e901\u7f16\u7801"
                    },
                    "source1Name": {
                        "type": "string",
                        "description": "\u6570\u636e\u6e901\u540d\u79f0"
                    },
                    "source2Code": {
                        "type": "string",
                        "description": "\u6570\u636e\u6e902\u7f16\u7801"
                    },
                    "source2Name": {
                        "type": "string",
                        "description": "\u6570\u636e\u6e902\u540d\u79f0"
                    },
                    "source3Code": {
                        "type": "string",
                        "description": "\u6570\u636e\u6e903\u7f16\u7801"
                    },
                    "source3Name": {
                        "type": "string",
                        "description": "\u6570\u636e\u6e903\u540d\u79f0"
                    },
                    "source4Code": {
                        "type": "string",
                        "description": "\u6570\u636e\u6e904\u7f16\u7801"
                    },
                    "source4Name": {
                        "type": "string",
                        "description": "\u6570\u636e\u6e904\u540d\u79f0"
                    },
                    "source5Code": {
                        "type": "string",
                        "description": "\u6570\u636e\u6e905\u7f16\u7801"
                    },
                    "source5Name": {
                        "type": "string",
                        "description": "\u6570\u636e\u6e905\u540d\u79f0"
                    },
                    "lawClauseCode": {
                        "type": "string",
                        "description": "\u6cd5\u6761\u7f16\u7801"
                    },
                    "remark": {
                        "type": "string",
                        "description": "\u5907\u6ce8"
                    },
                    "status": {
                        "type": "number",
                        "description": "\u72b6\u6001\uff1a1-\u542f\u7528\uff0c0-\u505c\u7528"
                    }
                }
            },
...
        }
    }
}